test(protocol): refactor helper test methods move generateFactoryID() to a Factory.js helper test file remove redundant note.js test helper method test file
