+++ packages/protocol/test/ACE/NoteRegistry/NoteRegistryManager.js
@@ -10,7 +10,8 @@ const secp256k1 = require('@aztec/secp256k1');
 
 // ### Internal Dependencies
 /* eslint-disable-next-line object-curly-newline */
-const { getNotesForAccount } = require('../../helpers/note');
+const factoryHelpers = require('../../helpers/Factory');
+const { getNotesForAccount } = require('../../helpers/ERC1724');
 
 // ### Artifacts
 const ACE = artifacts.require('./ACE');
@@ -21,10 +22,6 @@ const JoinSplitValidator = artifacts.require('./JoinSplit');
 const TestFactory = artifacts.require('./test/TestFactory');
 
 const { JOIN_SPLIT_PROOF } = proofs;
-
-const generateFactoryId = (epoch, cryptoSystem, assetType) => {
-    return epoch * 256 ** 2 + cryptoSystem * 256 ** 1 + assetType * 256 ** 0;
-};
 
 contract('NoteRegistryManager', (accounts) => {
     const [owner, notOwner, zkAssetOwner] = accounts;
@@ -43,7 +40,9 @@ contract('NoteRegistryManager', (accounts) => {
         const epoch = 1;
         const cryptoSystem = 1;
         const assetType = 0b01; // (adjust, canConvert) in binary;
-        await ace.setFactory(generateFactoryId(epoch, cryptoSystem, assetType), factoryContract.address, { from: owner });
+        await ace.setFactory(factoryHelpers.generateFactoryId(epoch, cryptoSystem, assetType), factoryContract.address, {
+            from: owner,
+        });
         erc20 = await ERC20Mintable.new();
         await erc20.mint(zkAssetOwner, 100);
         await erc20.approve(ace.address, 100, { from: zkAssetOwner });
@@ -51,7 +50,9 @@ contract('NoteRegistryManager', (accounts) => {
 
     describe('Success States', async () => {
         it('should register a factory', async () => {
-            const receipt = await ace.setFactory(generateFactoryId(1, 2, 1), factoryContract.address, { from: owner });
+            const receipt = await ace.setFactory(factoryHelpers.generateFactoryId(1, 2, 1), factoryContract.address, {
+                from: owner,
+            });
             const event = receipt.logs.find((l) => l.event === 'SetFactory');
             expect(event.args.epoch.toNumber()).to.equal(1);
             expect(event.args.cryptoSystem.toNumber()).to.equal(2);
@@ -73,7 +74,7 @@ contract('NoteRegistryManager', (accounts) => {
         });
 
         it('should deploy using the most recent factory', async () => {
-            const factoryAddress = await ace.getFactoryAddress(generateFactoryId(1, 1, 1));
+            const factoryAddress = await ace.getFactoryAddress(factoryHelpers.generateFactoryId(1, 1, 1));
             const canAdjustSupply = false;
             const canConvert = true;
             await ace.createNoteRegistry(erc20.address, scalingFactor, canAdjustSupply, canConvert, {
@@ -102,7 +103,7 @@ contract('NoteRegistryManager', (accounts) => {
             const canConvert = true;
             await ace.createNoteRegistry(erc20.address, scalingFactor, canAdjustSupply, canConvert, { from: zkAssetOwner });
             const existingProxy = await ace.registries(zkAssetOwner);
-            const newFactoryId = generateFactoryId(1, 3, 1);
+            const newFactoryId = factoryHelpers.generateFactoryId(1, 3, 1);
             const receipt = await ace.setFactory(newFactoryId, factoryContract.address, { from: owner });
 
             const { factoryAddress } = receipt.logs.find((l) => l.event === 'SetFactory');
@@ -141,7 +142,7 @@ contract('NoteRegistryManager', (accounts) => {
             const canConvert = true;
             await ace.createNoteRegistry(erc20.address, scalingFactor, canAdjustSupply, canConvert, { from: zkAssetOwner });
             const existingProxy = await ace.registries(zkAssetOwner);
-            const newFactoryId = generateFactoryId(1, 3, 1);
+            const newFactoryId = factoryHelpers.generateFactoryId(1, 3, 1);
             const newFactoryContract = await TestFactory.new(ace.address);
 
             await ace.setFactory(newFactoryId, newFactoryContract.address, { from: owner });
@@ -170,7 +171,7 @@ contract('NoteRegistryManager', (accounts) => {
             await ace.validateProof(JOIN_SPLIT_PROOF, zkAssetOwner, data, { from: zkAssetOwner });
             await ace.updateNoteRegistry(JOIN_SPLIT_PROOF, depositProof.eth.output, zkAssetOwner, { from: zkAssetOwner });
 
-            const newFactoryId = generateFactoryId(1, 3, 1);
+            const newFactoryId = factoryHelpers.generateFactoryId(1, 3, 1);
             await ace.setFactory(newFactoryId, factoryContract.address, { from: owner });
 
             await ace.upgradeNoteRegistry(newFactoryId, { from: zkAssetOwner });
@@ -181,13 +182,15 @@ contract('NoteRegistryManager', (accounts) => {
 
     describe('Failure States', async () => {
         it('should not register a factory if sent by non-owner', async () => {
-            await truffleAssert.reverts(ace.setFactory(generateFactoryId(1, 1, 1), factoryContract.address, { from: notOwner }));
+            await truffleAssert.reverts(
+                ace.setFactory(factoryHelpers.generateFactoryId(1, 1, 1), factoryContract.address, { from: notOwner }),
+            );
         });
 
         it('should not deploy if no factory exists', async () => {
             const canAdjustSupply = false;
             const canConvert = true;
-            const factoryId = generateFactoryId(1, 3, 1);
+            const factoryId = factoryHelpers.generateFactoryId(1, 3, 1);
             await truffleAssert.reverts(
                 ace.createNoteRegistry(erc20.address, scalingFactor, canAdjustSupply, canConvert, factoryId),
                 'expected the factory address to exist',
@@ -198,7 +201,7 @@ contract('NoteRegistryManager', (accounts) => {
             const epoch = 1;
             const cryptoSystem = 1;
             const assetType = 0b00; // (canAdjust, canConvert) in binary;
-            const factoryId = generateFactoryId(epoch, cryptoSystem, assetType);
+            const factoryId = factoryHelpers.generateFactoryId(epoch, cryptoSystem, assetType);
             await ace.setFactory(factoryId, factoryContract.address, { from: owner });
 
             const canAdjustSupply = true;
@@ -215,7 +218,7 @@ contract('NoteRegistryManager', (accounts) => {
             const epoch = 1;
             const cryptoSystem = 1;
             const assetType = 0b00; // (canAdjust, canConvert) in binary;
-            const newFactoryId = generateFactoryId(epoch, cryptoSystem, assetType);
+            const newFactoryId = factoryHelpers.generateFactoryId(epoch, cryptoSystem, assetType);
 
             await ace.createNoteRegistry(erc20.address, scalingFactor, canAdjustSupply, canConvert, { from: zkAssetOwner });
 
@@ -231,7 +234,7 @@ contract('NoteRegistryManager', (accounts) => {
             const epoch = 0;
             const cryptoSystem = 1;
             const assetType = 0b01; // (canAdjust, canConvert) in binary;
-            const newFactoryId = generateFactoryId(epoch, cryptoSystem, assetType);
+            const newFactoryId = factoryHelpers.generateFactoryId(epoch, cryptoSystem, assetType);
 
             await ace.createNoteRegistry(erc20.address, scalingFactor, canAdjustSupply, canConvert, { from: zkAssetOwner });
